Added Data Parser
Moved all data analysis to "data_parser".
The code is a bit sloppy right now because I've been pretty busy this week, but the graph shows the comparative speeds of the algorithms quite well. It also shows that the algorithm speed scales linearly with time allotted for executions which is what I was hoping for.