Cleaned up data parser
Reorganized how data is exported and parsed.
The data is now graphed nicely on a loglog plot (although I still haven't  figured out how to make a trendline fit nicely in the loglog plot).
Ready to create a bar graph from the data of the functions when they had 1 second to run.